requirements.txt
python-dotenv~=0.15.0
flask~=1.1.2
flask_sqlalchemy~=2.4.4
flask_marshmallow~=0.14.0
flask_cors~=3.0.9
flask_restplus~=0.13.0
flask_migrate~=2.5.3
marshmallow~=3.9.1
marshmallow_sqlalchemy~=0.24.1
sqlalchemy~=1.3.20
datetime~=4.3
gunicorn~=20.0.4
psycopg2-binary~=2.8.5
werkzeug==0.16.1
click~=7.1.2
alembic~=1.4.2
ezodf~=0.3.2
requests~=2.25.0
lxml~=4.6.2
scikit-learn~=0.24.0
pandas~=1.2.0
numpy~=1.19.5
pytest~=6.2.5
